requirements.txt
python-telegram-bot
pyrogram
tgcrypto
pymongo[srv]
python-dotenv